Moving the RecordReferenceEnhancer to the same directory as implementation
