+++ lib/Indexer/Model/RecordReferenceEnhancer/NullRecordReferenceEnhancer.php
@@ -3,7 +3,6 @@
 namespace Phpactor\Indexer\Model\RecordReferenceEnhancer;
 
 use Phpactor\Indexer\Model\RecordReference;
-use Phpactor\Indexer\Model\RecordReferenceEnhancer;
 use Phpactor\Indexer\Model\Record\FileRecord;
 
 class NullRecordReferenceEnhancer implements RecordReferenceEnhancer
